+++ react/components/LoginOptions.js
@@ -15,31 +15,22 @@ import OAuth from './OAuth'
 import styles from '../styles.css'
 
 const PROVIDERS_ICONS = {
-  'Google': GoogleIcon,
-  'Facebook': FacebookIcon,
+  Google: GoogleIcon,
+  Facebook: FacebookIcon,
 }
 
 /** LoginOptions tab component. Displays a list of login options */
 class LoginOptions extends Component {
-  state = {
-    loadingOptions: false,
-  }
-
-  handleRefetchOptions = () => {
-    if (!this.state.loadingOptions) {
-      this.setState({ loadingOptions: true })
-      this.props.refetchOptions()
-        .then(() => this.setState({ loadingOptions: false }))
-    }
-  }
-
   handleOptionClick = el => () => {
     this.props.onOptionsClick(el)
   }
 
   showOption = (option, optionName) => {
     const { isAlwaysShown, currentStep, options } = this.props
-    return options && ((options[option] && !isAlwaysShown) || currentStep !== optionName)
+    return (
+      options &&
+      ((options[option] && !isAlwaysShown) || currentStep !== optionName)
+    )
   }
 
   render() {
@@ -54,8 +45,6 @@ class LoginOptions extends Component {
       providerPasswordButtonLabel,
     } = this.props
 
-    const { loadingOptions } = this.state
-
     const classes = classNames(styles.options, className, {
       [styles.optionsSticky]: isAlwaysShown,
     })
@@ -64,64 +53,80 @@ class LoginOptions extends Component {
       <div className={classes}>
         <FormTitle>{title || translate(fallbackTitle, intl)}</FormTitle>
         <ul className={`${styles.optionsList} list pa0`}>
-          {options ? (
-            <Fragment>
-              {options.accessKeyAuthentication && this.showOption('accessKeyAuthentication', 'store/loginOptions.emailVerification') &&
-                <li className={`${styles.optionsListItem} mb3`}>
-                  <div className={styles.button}>
-                    <Button
-                      variation="secondary"
-                      onClick={this.handleOptionClick('store/loginOptions.emailVerification')}
-                    >
-                      <span>{translate('store/loginOptions.emailVerification', intl)}</span>
-                    </Button>
-                  </div>
-                </li>
-              }
-              {options.classicAuthentication && this.showOption('classicAuthentication', 'store/loginOptions.emailAndPassword') &&
-                <li className={`${styles.optionsListItem} mb3`}>
-                  <div className={styles.button}>
-                    <Button
-                      variation="secondary"
-                      onClick={this.handleOptionClick('store/loginOptions.emailAndPassword')}
-                    >
-                      <span>{providerPasswordButtonLabel || translate('store/loginOptions.emailAndPassword', intl)}</span>
-                    </Button>
-                  </div>
-                </li>
-              }
-              {options.providers && options.providers.map(({ providerName }, index) => {
+          <Fragment>
+            {options.accessKeyAuthentication &&
+              this.showOption(
+                'accessKeyAuthentication',
+                'store/loginOptions.emailVerification'
+              ) && (
+              <li className={`${styles.optionsListItem} mb3`}>
+                <div className={styles.button}>
+                  <Button
+                    variation="secondary"
+                    onClick={this.handleOptionClick(
+                      'store/loginOptions.emailVerification'
+                    )}
+                  >
+                    <span>
+                      {translate(
+                        'store/loginOptions.emailVerification',
+                        intl
+                      )}
+                    </span>
+                  </Button>
+                </div>
+              </li>
+            )}
+            {options.classicAuthentication &&
+              this.showOption(
+                'classicAuthentication',
+                'store/loginOptions.emailAndPassword'
+              ) && (
+              <li className={`${styles.optionsListItem} mb3`}>
+                <div className={styles.button}>
+                  <Button
+                    variation="secondary"
+                    onClick={this.handleOptionClick(
+                      'store/loginOptions.emailAndPassword'
+                    )}
+                  >
+                    <span>
+                      {providerPasswordButtonLabel ||
+                          translate(
+                            'store/loginOptions.emailAndPassword',
+                            intl
+                          )}
+                    </span>
+                  </Button>
+                </div>
+              </li>
+            )}
+            {options.providers &&
+              options.providers.map(({ providerName }, index) => {
                 const hasIcon = PROVIDERS_ICONS.hasOwnProperty(providerName)
                 return (
                   <li
                     className={`${styles.optionsListItem} mb3`}
                     key={`${providerName}-${index}`}
                   >
-                    <OAuth provider={providerName} loginCallback={loginCallback}>
-                      {hasIcon ? React.createElement(PROVIDERS_ICONS[providerName], null) : null}
+                    <OAuth
+                      provider={providerName}
+                      loginCallback={loginCallback}
+                    >
+                      {hasIcon
+                        ? React.createElement(
+                          PROVIDERS_ICONS[providerName],
+                          null
+                        )
+                        : null}
                     </OAuth>
                   </li>
                 )
               })}
-            </Fragment>
-          ) : (
-            <li className={`${styles.optionsListItem} mb3`}>
-              <div className={`${styles.button} ${styles.buttonDanger}`}>
-                <Button
-                  type="danger"
-                  variation="secondary"
-                  isLoading={loadingOptions}
-                  onClick={this.handleRefetchOptions}
-                >
-                  <div>
-                    <div className={`${loadingOptions ? 'dn' : 'db'}`}>{translate('store/loginOptions.error.title', intl)}</div>
-                    <div className="t-small pt1">{translate('store/loginOptions.error.subhead', intl)}</div>
-                  </div>
-                </Button>
-              </div>
-            </li>
-          )}
-          <li className={`${styles.optionsListItem} ${styles.optionsListItemContainer} mb3`}>
+          </Fragment>
+          <li
+            className={`${styles.optionsListItem} ${styles.optionsListItemContainer} mb3`}
+          >
             <ExtensionContainer id="container" />
           </li>
         </ul>
@@ -135,8 +140,6 @@ LoginOptions.propTypes = {
   intl: intlShape,
   /** Function to change de active tab */
   onOptionsClick: PropTypes.func.isRequired,
-  /** Function to refetch login options */
-  refetchOptions: PropTypes.func.isRequired,
   /** Title that will be shown on top */
   title: PropTypes.string.isRequired,
   /** Fallback title that will be shown if there's no title */
@@ -145,10 +148,12 @@ LoginOptions.propTypes = {
   options: PropTypes.shape({
     accessKeyAuthentication: PropTypes.bool,
     classicAuthentication: PropTypes.bool,
-    providers: PropTypes.arrayOf(PropTypes.shape({
-      className: PropTypes.string,
-      providerName: PropTypes.string,
-    })),
+    providers: PropTypes.arrayOf(
+      PropTypes.shape({
+        className: PropTypes.string,
+        providerName: PropTypes.string,
+      })
+    ),
   }),
   /** Class of the root element */
   className: PropTypes.string,